Test dependency isolation between crates
A nix flake that follows the basic best practices outlined by crane will build external dependencies separately from local crates. This is typically configured as single workspace-wide step, and is often appropriate: an external dependency used by 3 local crates will only be compiled once. However, there are scenarios where the overlap of dependencies between the local crates is smaller or less complex than exclusive dependencies. In this case, it can be preferable to build dependencies separately for each crate.

This amends the test script with failing checks for this optimization.